# docs/swagger.yaml
basePath: /
definitions:
...
  gear.Gear:
    properties:
      created_at:
        type: string
      date_begin:
        type: string
      date_end:
        type: string
      display_name:
        type: string
      external_id:
        type: string
      gear_type:
        $ref: '#/definitions/gear.Type'
      id:
        type: string
      retired:
        type: boolean
      total_activities:
        type: integer
      total_distance_m:
        type: number
      updated_at:
        type: string
    type: object
  gear.Type:
    enum:
    - shoes
    - bike
    - other
    type: string
    x-enum-varnames:
    - TypeShoes
    - TypeBike
    - TypeOther
  gear.upsertRequest:
    properties:
      date_begin:
        type: string
      date_end:
        type: string
      display_name:
        type: string
      external_id:
        type: string
      gear_type:
        type: string
      retired:
        type: boolean
      total_activities:
        type: integer
      total_distance_m:
        type: number
    type: object
...
  hydration.Daily:
    properties:
      date:
        type: string
      entries:
        items:
          $ref: '#/definitions/hydration.Entry'
        type: array
      entry_count:
        type: integer
      total_ml:
        type: number
      tz:
        type: string
    type: object
  hydration.Entry:
    properties:
      created_at:
        type: string
      id:
        type: string
      logged_at:
        type: string
      note:
        type: string
      quantity_ml:
        type: number
      updated_at:
        type: string
      workout_id:
        type: string
    type: object
  hydration.createRequest:
    properties:
      logged_at:
        type: string
      note:
        type: string
      quantity_ml:
        type: number
      workout_id:
        type: string
    type: object
  hydration.patchRequest:
    properties:
      logged_at:
        type: string
      note:
        type: string
      quantity_ml:
        type: number
      workout_id:
        description: |-
          WorkoutID supports the empty-string clear sentinel:
            omitted   → leave unchanged
            "<uuid>"  → set the link
            ""        → clear the link
        type: string
    type: object
...
info:
  contact: {}
  description: Kazper — personal endurance-fueling and training-coach REST API. All
    endpoints under /products, /meals, and /summary require a bearer token.
  title: Kazper
  version: 0.1.0
paths:
...
  /gear:
    get:
      parameters:
      - description: Filter by retirement state (true|false)
        in: query
        name: retired
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: '{ gear: [...] }'
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List gear records
      tags:
      - gear
    post:
      consumes:
      - application/json
      description: Creates or updates a piece of gear, upserting by `external_id`
        (the Garmin gear uuid). Slowly-changing inventory — re-observing the same
        gear updates it in place. Standard `Idempotency-Key` header supported. Gear
        is coaching context only; its distance never feeds nutrition computation.
      parameters:
      - description: Optional client-supplied idempotency key
        in: header
        name: Idempotency-Key
        type: string
      - description: Gear record
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/gear.upsertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: UPDATE (external_id already present)
          schema:
            $ref: '#/definitions/gear.Gear'
        "201":
          description: INSERT
          schema:
            $ref: '#/definitions/gear.Gear'
        "400":
          description: external_id_required | gear_type_invalid | display_name_required
            | total_distance_m_invalid | total_activities_invalid
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upsert a gear record (by Garmin gear id)
      tags:
      - gear
  /gear/{id}:
    get:
      parameters:
      - description: Gear UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gear.Gear'
        "404":
          description: gear_not_found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a gear record by id
      tags:
      - gear
...
  /hydration:
    get:
      parameters:
      - description: Inclusive RFC3339 lower bound
        in: query
        name: from
        required: true
        type: string
      - description: Exclusive RFC3339 upper bound
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{ entries: [...] }'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: window_required | window_invalid | range_too_large
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List hydration entries in a window
      tags:
      - hydration
    post:
      consumes:
      - application/json
      description: Records a volume of fluid drunk at a moment in time. The optional
        `note` carries beverage context (e.g. "water", "coffee", "electrolytes").
      parameters:
      - description: Optional client-supplied idempotency key
        in: header
        name: Idempotency-Key
        type: string
      - description: Hydration entry
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/hydration.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hydration.Entry'
        "400":
          description: quantity_ml_invalid | logged_at_invalid | logged_at_too_far_future
            | note_too_long
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Log a hydration entry
      tags:
      - hydration
...
  /hydration/{id}:
    delete:
      parameters:
      - description: Hydration entry UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: no content
        "404":
          description: hydration_not_found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a hydration entry
      tags:
      - hydration
    patch:
      consumes:
      - application/json
      parameters:
      - description: Hydration entry UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/hydration.patchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hydration.Entry'
        "400":
          description: quantity_ml_invalid | logged_at_invalid | note_too_long
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: hydration_not_found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Partially update a hydration entry
      tags:
      - hydration
...
  /summary/hydration/daily:
    get:
      description: Returns total ml and per-entry list for one calendar date in the
        supplied timezone. Volume-only — separate from /summary/daily, which is nutrient-only.
      parameters:
      - description: Calendar date in YYYY-MM-DD
        in: query
        name: date
        required: true
        type: string
      - description: IANA timezone (defaults to DEFAULT_USER_TZ)
        in: query
        name: tz
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hydration.Daily'
        "400":
          description: date_invalid | tz_invalid
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Daily hydration total + entries
      tags:
      - summary
...
securityDefinitions:
  BearerAuth:
    description: Static bearer token. Use the format `Bearer <token>` with either
      MOBILE_API_TOKEN or AGENT_API_TOKEN.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"